==> R-scripts/08-RGEE_TomVersluijs_S2_Shapefile_Points_Snowmelt.R <==
#######################################################################################################################################

#The timing of snow melt is calculated based on Sentinel-2 data for all point locations located within a single polygon. 
#The user can specify a buffer zone (radius) to depict the area in which snow melt will be analysed per location. All 
#locations (including buffer zone) are required to be located within a single shapefile and are then analysed simultaneously. 
#First, clouds and permanent water bodies are filtered within the shapefile. Second, if the shapefile overlaps with multiple 
#satellite tiles for a certain day, a composite image is created (picking the pixel with least cloud cover). Finally, snow melt 
#is analysed for each location's buffer zone based on one of the following methods (specified by the user by setting the 
#parameter 'method'):

# (I):  'avg_NDSI':     Calculate the average NDSI value over time within each point's buffer zone, fits a GAM through these 
#                       data and calculates when this model passes the specified NDSI threshold defining the moment of snow
#                       melt. In addition, time series of the average NDVI and NDMI are extracted within each point's buffer
#                       zone.
# (II): 'snowfraction': Calculate the fraction of pixels within each buffer zone over time where NDSI > 'NDSI_threshold', fits 
#                       a GAM through these data and extract the moment when this model passes a user-specified 'Snowfraction_
#                       threshold'.

#Note that snow melt is not calculated based on pixel-level GAM fits (i.e. 'pixel_gam' method is not implemented). This approach 
#is instead implemented in script "06-RGEE_TomVersluijs_S2_Points_Snowmelt.R", and in script "05-RGEE_TomVersluijs_S2_Pixels_
#Snowmelt.R". In the former script, these calculations are made for all pixels within the buffer zone of point locations. In the 
#latter script these calculates are made for all pixels within a single polygon, resulting in a pixel-level map of the timing of 
#snow melt. Script "10-RGEE_TomVersluijs_ExtractSnowFraction.R" can then be used to extract time series of the fraction of snow 
#cover for points/polygons of interest from this map.

#The current script (08) is similar to the script '06-RGEE_TomVersluijs_S2_Points_Snowmelt.R'. However, in the latter script all 
#points are analysed consecutively using a loop, which makes that script significantly slower to run. That script does not rely 
#on a shapefile and thus works for points spaced much further apart (i.e. tracking data of migratory birds). Script 08 only works 
#for small areas of c.a. 50-100 km2 (larger areas might result in computation errors unless the spatial resolution of the analyses 
#is decreased). In addition, the 'pixel_gam' method is implemented in script "06", but not in the current script "08".

#Copyright Tom Versluijs 2026-02-19. Do not use this code without permission. Contact information: tom.versluijs@gmail.com

#Before running this script make sure to install RGEE according to the instructions in script "00-RGEE_TomVersluijs_Installation.R". 
#Note that a GoogleDrive is required. Important: make sure to run this script from within the "RGEE_Snowmelt.Rproj" project file.

##################################################################################################################################

#I: Setup workspace

##################################################################################################################################

      #(0): Setup workspace

       #Clear workspace
       rm(list=ls())
       utils::install.packages("here")
       library(here)
       
       #Load python environment path
       if(file.exists(paste0(here::here(), "/Input/rgee_environment_dir.rds"))){
         rgee_environment_dir <- readRDS(paste0(here::here(), "/Input/rgee_environment_dir.rds"))
         reticulate::use_python(rgee_environment_dir, required=T)
         reticulate::py_config()
         }
       
       #Load google earth engine credentials folder
       if(file.exists(paste0(here::here(), "/Input/rgee_credentials_dir.rds"))){
         fldr_credentials_user <- readRDS(paste0(here::here(), "/Input/rgee_credentials_dir.rds"))
         }
        
      #(1): Load packages
       #renv::restore() #revert to last version of R-packages used to successfully run this script (optional).
       utils::install.packages("pacman")
       library(pacman)
       p_load(sf, rgee, ggplot2, mgcv, googledrive, dplyr, tidyr, foreach, parallel, doSNOW, gridExtra, rgeeExtra, magick)

      #(2): Define ggplot2 plotting theme
        theme_tom <- function(){
           theme_classic() %+replace%
             theme(axis.title = element_text(size=18),
                   axis.text = element_text(size=16),
                   legend.position = "none",
                   strip.background = element_rect(fill = "white", colour = "black", linewidth = rel(2)),
                   complete = TRUE)}    
        
      #(3): Load auxiliary functions
        source_files <- list.files(path=paste0(here(), "/Input"), full.names=TRUE, recursive = TRUE, pattern = "Sentinel2_AuxiliaryFunctions|Installation")
        sapply(source_files, source, chdir = TRUE) ; rm(source_files)
        
      #(4): Initialize earth engine using a custom function
        f_ee_Init(user = "tom.versluijs@gmail.com", project = "heroic-bird-255915", fldr_asset = "escape", 
                  fldr_credentials_user=fldr_credentials_user, drive = TRUE, quiet = FALSE)
        #rgee::ee_Initialize(user = "tom.versluijs@gmail.com", drive = TRUE)
        
##################################################################################################################################
       
#II: Specify parameters of interest
       
##################################################################################################################################       
       
#(4): Manually specify parameters of interest

  #(a): Sentinel-2 satellite

    #Sentinel-2 dataset ("S2_HARMONIZED" or "S2_SR_HARMONIZED")
    s2_dataset <- "S2_SR_HARMONIZED"
    #Use S2_HARMONIZED (level 1c product) for data from 2016 - 2018
    #Use S2_SR_HARMONIZED (level 2a product) for data starting from 2019

    #Resolution of sampling in meters
    resolution=20 #default maximum resolution for Sentinel-2 = 10m

  #(b) Area of interest

    #Specify name of study area (used as prefix in output files)
    area_name="ZAC" #max length three characters

    #Name of Shapefile (located in '/Input/Shapefiles' folder)
	  #Follow the guide "Manual_CreateShapefilePolygons.docx" when creating this shapefile.
    shapefile <- "ZAC_Outline_EPSG4326.shp"

    #Coordinate reference system used for calculations
    #EPSG:4326 is recommended for areas spanning multiple UTM zones, but increases computation time (i.e. spherical coordinate system).
    #EPSG:326XX might result in reduced computation time for areas located within a single UTM zone (i.e. planar coordinate system).
    crs <- "EPSG:4326"

  #(c) Point locations

    #Name of file with point locations located in Input folder
    input_locations <- "TestLocations_Zackenberg.csv"
    #Make sure it contains the coordinates in decimal degrees in the columns "LON_x" and "LAT_y"
   
    #Buffer radius around each point location (in meters)
    Buffer_radius_m=250

  #(c) Dates

    #Specify the year of interest:
    year_ID <- "2022"

    #Date range of all images considered for analysis
    start_date <- paste0(year_ID, "-03-15") #choose date (well) before the first snow melt occurs within the study area (default -03-15)
    end_date <- paste0(year_ID, "-09-15") #choose date (well) after last snow melt occurs within the study area (default -09-15)

  #(d) Snow detection

    #NDSI threshold for snow detection (specify multiple using c())
    NDSI_threshold_vector = c(0.4, 0.3, 0.5)

    #Define the snowcover fraction for which the date of its occurrence will be calculated (specify multiple using c())
    Snowfraction_threshold_vector = c(0.25, 0.5, 0.75)
    
    #Define the preferred method for the analysis of snowmelt
    method=c("avg_NDSI", "snowfraction") #either "avg_NDSI", "snowfraction" or both using c()
    #(1) "avg_NDSI":     Calculate the average NDSI, NDVI and NDMI values within each point's buffer zone over time and extracts 
    #                    the moment when the NDSI value is equal to 'NDSI_threshold'.
    #(2) "snowfraction": Count the fraction of pixels within each point's buffer zone over time where NDSI > 'NDSI_threshold' 
    #                    and extract the moment when this fraction is equal to 'Snowfraction_threshold'.
        
  #(e): Cloud masking

    #Should clouds be masked from the analysis (default=TRUE).
    #Note that cloud masking will automatically be set to TRUE when mask_water=TRUE
    #Highly recommended to set to TRUE.
    mask_clouds=TRUE

    #Cloud probability threshold (ranging 0-100) for S2_CLOUD_PROBABILITY dataset
    #Defines above which value of the cloud probability parameter a pixel is considered cloudy. Higher values
    #are thus more conservative (i.e. pixels are less likely to be classified as cloudy)
    cldprb_threshold=70

    #Define radius of cloud erosion and dilation (in meters)
    #First, all cloudy pixels without cloudy pixels within radius cld_erosion will be removed.
    #Second, all remaining pixels will be dilated by radius cld_buffer. Make sure that cld_buffer
    #is larger than cld_erosion. Larger values result in a more course cloud-filtering process.
    cld_erosion=250 #250 for conservative cloud estimation
    cld_buffer=270 #270 for conservative cloud estimation

    #Define resolution of cloud-filtering (in meters)
    resolution_cldmsk=150 #(default = 150)

    #Maximum fraction of cloud cover allowed in each image
    max_cloud_fraction=0.75 #0.75 equals 75% cloud cover

  #(f): Water masking

    #Should permanent waterbodies be masked from the analysis (default=TRUE).
    mask_water=TRUE
    mask_water_type="water_mask_ESA" #either "water_mask_ESA", "water_mask_Manual", or "both"
    #"water_mask_ESA" uses the world wide ESA dataset with a 10m resolution which works well with large areas (default)
    #"water_mask_Manual" uses a manual approach based on NDWI, NDSI and NIR bands and generally works well for small details.
    #"both" employs both methods and sets pixels to water if one or both of these methods indicates so.

    #Parameters for type "water_mask_Manual"

      #NDWI threshold above which is pixel is perceived as water (or snow!)
      NDWI_threshold=0.0

      #NDSI and NIR thresholds above which a pixel is perceived as snow
      NIR_threshold=1100

      #Date range to detect permanent waterbodies based on NDWI
      start_date_NDWI <- paste0(year_ID, "-07-15")
      end_date_NDWI <- paste0(year_ID, "-08-15")

  #(g): Composite image
      
    #Specify whether a composite image will be generated for all images (i.e. tiles) per unique doy (default=TRUE). Note that setting this
    #to TRUE might result in an increase in computation time.
    create_composite=TRUE
    
  #(h): GAM sequential outlier filtering

    #parameters for auxiliary function to filter outliers from GAM using a two-step approach
    outlier_removal=TRUE #should sequential outlier removal be employed when fitting GAMs to the data
    outlier_thresh_1=0.4 #first threshold in outlier removal (relative to y-range of data, default=0.4)
    outlier_thresh_2=0.2 #second threshold in outlier removal (relative to y-range of data, default=0.2)
      
    #Specify the degree of GAM smoothing by setting the 'k' parameter
    gam_k_outlier=10 #Number of knots when filtering outliers (default=10)
    gam_k=25 #Number of knots when making model predictions (default=25). 
    #Larger values result in a more precise GAM-fit, but might result in overfitting.
       
  #(i): Counts of unmasked pixels
    
    #Should counts of the number of unmasked pixels per day of year within the shapefile area be conducted (increases computation time)
    pixel_counts=TRUE
    
  #(j): GIF animations
    
    #Should GIF animations be constructed (increases computation time)
    gif_output=FALSE
    
    #Maximum dimension (in pixels) of GIF
    gif_max_pixels=250
    
##################################################################################################################################

#III: Define some additional parameters (automated)

##################################################################################################################################
    
  #(5): Automatically define some additional parameters 
    
    #Check if date variables are specified correctly
     if(!(as.Date(start_date) <= (as.Date(start_date_NDWI) - as.difftime(14, units = "days")))){
       stop("'start_date' should be set at least 14 days earlier than 'start_date_NDWI'")
       }
     if(!(as.Date(end_date) >= as.Date(end_date_NDWI))){
       stop("'end_date' should be set equal to or later than 'end_date_NDWI'")
       }
    
    #Create a unique data_ID
    if(nchar(area_name)>3){area_name <- substr(area_name, start = 1, stop = 3)}
    data_ID <- paste0(area_name, substr(year_ID,(nchar(year_ID)+1)-2,nchar(year_ID)), "_S2")           
    
    #Create a timestamp variable
    timestamp <- format(Sys.time(), "%Y%m%d%H%M%S")
    
    #Specify starting and ending date as day of year 
    start_date_doy <- as.numeric(strftime(start_date, format = "%j"))
    end_date_doy <- as.numeric(strftime(end_date, format = "%j"))   
    
    #Set mask_clouds to TRUE if a composite image needs to be generated         
    if(create_composite==TRUE){mask_clouds=TRUE}      
    
    #Set mask_clouds to TRUE if water masking is TRUE and mask_water_type==water_mask_Manual
    if(mask_water==TRUE & (mask_water_type=="water_mask_Manual" | mask_water_type=="both")){mask_clouds=TRUE}      
    
    #Create output folder
    if(dir.exists(paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt"))==FALSE){dir.create(paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt"), recursive = TRUE)}
    
    #Save all parameters and their values in the environment to a text file 
    file_conn <- file(paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Parameters.txt"), "w")
    for (obj in setdiff(ls(), lsf.str())) {cat(paste(obj, "=", get(obj)), file = file_conn) ; cat("\n", file = file_conn)}
    close(file_conn)
    
##################################################################################################################################

#III: Read and display the unfiltered data

##################################################################################################################################

    #(6) Read aoi_Shapefile and convert to a feature collection
      root_fldr <- here()
      aoi_Shapefile <- st_read(paste0(root_fldr, "/Input/Shapefiles/", shapefile), quiet=T)
      aoi_Shapefile <- st_transform(aoi_Shapefile, crs="EPSG:4326")
      aoi_Shapefile <- sf_as_ee(aoi_Shapefile)
      aoi_Shapefile <- ee$FeatureCollection(aoi_Shapefile)
      #This feature collection can be used to clip the satellite imagery to the study area
      
       #Calculate bounding box for aoi_Shapefile
        aoi <- aoi_Shapefile$geometry()$bounds()
        
       #Calculate central point
        coordinates_point <- aoi_Shapefile$geometry()$centroid()
        
       #Plot shapefile and bounding box (for debugging)
        Map$setCenter(coordinates_point$getInfo()$coordinates[1], coordinates_point$getInfo()$coordinates[2], 6)
        Map$addLayer(aoi_Shapefile)+
          Map$addLayer(ee$FeatureCollection(aoi)$style(color='red', fillColor='00000000'))+
          Map$addLayer(coordinates_point, list(color="blue"))
        
       #Calculate the size of the study area in km2:
        img <- ee$Image$pixelArea()$divide(1000000)
        area2 <- img$reduceRegion(
          reducer= ee$Reducer$sum(),
          geometry= aoi_Shapefile,
          crs=crs,
          scale= resolution,
          maxPixels= 1E13)
        paste0('Size of study area of calculated using the pixel area method: ', round(ee$Number(area2$get('area'))$getInfo(),3), ' km2')
      
    #(7) Plot an RGB, NDSI and NDVI image for the first extracted satellite image

      #Extract Sentinel-2 Surface Reflectance data for a given area and date range:
      s2_col <- ee$ImageCollection(paste0('COPERNICUS/', s2_dataset))
      s2_col <- s2_col$
        filterBounds(aoi)$ #For areas that do not span multiple tiles, change aoi to coordinates_point to greatly speed-up code.
        filterDate(start_date, end_date)
  
      #Add Sentinel-2 cloud probability as a new property to s2_col
      if(mask_clouds==TRUE){
        
        #Load cloud information from S2_CLOUD_PROBABILITY
        s2_cloudless_col <-ee$ImageCollection('COPERNICUS/S2_CLOUD_PROBABILITY')
        s2_cloudless_col <- s2_cloudless_col$
          filterBounds(aoi)$ #For areas that do not span multiple tiles, change aoi to coordinates_point to greatly speed-up code.
          filterDate(start_date, end_date) 
      
        #Store sentinel-2 projection (of band "probability")
        s2_projection <- s2_cloudless_col$first()$select("probability")$projection() 
        
        #Add Sentinel-2 cloud probability as a new property to s2_col
        s2_col <- Add_CloudProb_Property(s2_col=s2_col, s2_cloudless_col=s2_cloudless_col)
        
        #Make sure that each image contains the s2cloudless property (create an empty property if it does not)
        s2_col <- s2_col$map(Add_NULL_s2cloudless)
        
        # #Check that s2cloudless was added as a property (for debugging)
        #  s2_col$first()$propertyNames()$getInfo() 
      
      }
      
      #Clip all images to the area depicted by aoi_Shapefile
      s2_col <- s2_col$map(function(img){return(img$clipToCollection(aoi_Shapefile))})
      
    #(8): Plot an RGB, NDSI and NDVI image for a single extracted satellite image (for debugging)
      
      # #Select a single image for initial plot:
      # image <- s2_col$filterDate(paste0(year_ID, "-07-15"), end_date)$first()
      # 
      # #Normalized difference snow index:
      # ndsi_s2 <-  image$expression('(B3 - B11) / (B3 + B11)',
      #                              list('B11'=image$select('B11'),
      #                                   'B3'=image$select('B3')))
      # 
      # #Normalized difference vegetation index:
      # ndvi_s2 <- image$expression('(B8 - B4) / (B8 + B4)',
      #                             list('B8'=image$select('B8'),
      #                                  'B4'=image$select('B4')))
      # 
      # #Add Normalized difference water index to this image:
      # ndwi_s2 <- image$expression('(B3 - B8) / (B3 + B8)',
      #                             list('B3'=image$select('B3'),
      #                                  'B8'=image$select('B8')))
      # 
      # #Plot all layers
      # Map$setCenter(coordinates_point$getInfo()$coordinates[1], coordinates_point$getInfo()$coordinates[2], 10)
      # Map$addLayer(image,list(bands=c("B4", "B3", "B2"), min=100, max=8000, gamma=c(1.9, 1.7, 1.7)), 'TRUE COLOR')+ 
      # Map$addLayer(ndsi_s2,list(min=-1, max=1.5, palette=c('black', '0dffff', '0524ff', 'ffffff')), 'NDSI')+
      # Map$addLayer(ndvi_s2,list(min=-1, max=1, palette=c('#FF0000','#00FF00')), 'NDVI')+
      # Map$addLayer(ndwi_s2,list(min=0, max=1, palette=c('000000', '0dffff', '0524ff', 'ffffff')), 'NDWI')
      
    #(9): Define a featurecollection of points for which Sentinel-2 data needs to be extracted
      ZAC_Paths <- read.csv(paste0("Input/", input_locations), header=T)[,c("LON_x", "LAT_y")]
      colnames(ZAC_Paths) <- c("LON_x", "LAT_y")
      
      #Add a unique LocationID to every unique lat/lon combination
      Locations_sf <- unique(ZAC_Paths[,c("LON_x", "LAT_y")])
      Locations_sf$LocationID <- paste0("Location_", 1:nrow(Locations_sf))
      ZAC_Paths <- left_join(ZAC_Paths, Locations_sf, by=c("LON_x", "LAT_y"))
      
      Locations  <-  st_as_sf(Locations_sf, coords = c("LON_x", "LAT_y"), crs = crs)
      Locations <- st_transform(Locations, crs=crs)
      Locations <- sf_as_ee(Locations[,c("LocationID", "geometry")])
      
      #Add buffer around all point locations
      if(Buffer_radius_m>0){
        bufferBy <- function(Buffer_radius_m) {
        return(function(feature) {
          return(feature$buffer(Buffer_radius_m))   
          })
        }
        Locations <- Locations$map(bufferBy(Buffer_radius_m))
        }
      
      #Locations is now a feature collection. A feature in our analysis comprises the feature type (point in case no buffer
      #was added, or a circular polygon with radius Buffer_radius_m with a center the original point location), Location_ID  
      #and the feature's geometry as lat and lon coordinates. A feature collection corresponds to a collection of such features 
      #(i.e. a collection of different Locations, with their respective feature types, LocationIDs, DateTimes and coordinates).
      
      #Plot feature collection on a map (i.e. locations as points)
      image <- s2_col$filterDate(paste0(year_ID, "-06-15"), end_date)$first()
      Map$setCenter(coordinates_point$getInfo()$coordinates[1], coordinates_point$getInfo()$coordinates[2], 10)
      Map$addLayer(aoi_Shapefile)+
      Map$addLayer(image,list(bands=c("B4", "B3", "B2"), min=0, max=12000, gamma=c(1.9, 1.7, 1.7)), 'TRUE COLOR')+
      Map$addLayer(Locations, list(color="red"), paste0("Locations_", year_ID))    
      
    #(10) Add NDSI, NDVI, NDMI and NDWI to the clipped image collection
      
     #Map over image collection
      s2_col <- s2_col$
        map(getNDSI)$
        map(getNDVI)$
        map(getNDMI)$
        map(getNDWI)$
        map(add_Date)
    
    #(11) Create a timeseries GIF of unmasked RGB and NDSI images for aoi_Shapefile
      
      if(gif_output==TRUE){
        
        #(a): GIF of RGB-images
        f_img_col_to_gif(img_col=s2_col,  #image collection
                         RGB_bands=c("B4", "B3", "B2"), #names of RGB-bands in image collection
                         shapefile=aoi_Shapefile, #shapefile of area of interest
                         centroid_buffer_m=0, #buffer zone surrounding centroid of shapefile in meters (to enlarge area for GIF)
                         gif_dimensions=gif_max_pixels, #Maximum dimensions of GIF (pixels)
                         gif_fps=5, #Frames per second of GIF
                         gif_min=0, #Value to map to 0 (to improve contrast). A good rule of thumb is to set min to values that represent the 2nd percentile of the data
                         gif_max=12000, #Value to map to 255 (to improve contrast). A good rule of thumb is to set max to values that represent the 98th percentile of the data
                         gif_gamma=c(1.9, 1.7, 1.7), #Gamma correction factors (one for each band)
                         gif_crs='EPSG:3857', #CRS project of the output
                         gif_text_position="northwest", #Location of text (datetime string)
                         gif_text_position_adjustment="+0+0", #Small scale adjustment of text in meters
                         gif_text_size=14, #Size of text
                         gif_text_col="#FFFFFF",
                         output_fldr=paste0(root_fldr, "/Output/S2/08_Shapefile_Points_Snowmelt/"),
                         file_name="GIF_RGB")
        
        #(b): GIF of NDSI-images
        
          #Transform NDSI-band to RGB bands
          s2_col_ndsi <- s2_col$map(ee_utils_pyfunc(function(image){
            f_band_to_RGB(img=image, band='NDSI', min_value=-1, max_value=1.5, palette=c('#000000', '#0dffff', '#0524ff', '#ffffff'))}))
          
          #Create gif of the transformed NDSI-band
          f_img_col_to_gif(img_col=s2_col_ndsi,  #image collection
                           RGB_bands=c("vis-red", "vis-green", "vis-blue"), #names of RGB-bands in image collection
                           shapefile=aoi_Shapefile, #shapefile of area of interest
                           centroid_buffer_m=0, #buffer zone surrounding centroid of shapefile in meters (to enlarge area for GIF)
                           gif_dimensions=gif_max_pixels, #Maximum dimensions of GIF (pixels)
                           gif_fps=5, #Frames per second of GIF
                           gif_min=0, #Value to map to 0 (to improve contrast). A good rule of thumb is to set min to values that represent the 2nd percentile of the data
                           gif_max=255, #Value to map to 255 (to improve contrast). A good rule of thumb is to set max to values that represent the 98th percentile of the data
                           gif_gamma=c(1, 1, 1), #Gamma correction factors (one for each band)
                           gif_crs='EPSG:3857', #CRS project of the output
                           gif_text_position="northwest", #Location of text (datetime string)
                           gif_text_position_adjustment="+0+0", #Small scale adjustment of text in meters
                           gif_text_size=14, #Size of text
                           gif_text_col="#FFFFFF",
                           output_fldr=paste0(root_fldr, "/Output/S2/08_Shapefile_Points_Snowmelt/"),
                           file_name="GIF_NDSI")
        
      }

##################################################################################################################################
        
#IV: Filter and mask clouds within the image collection
        
##################################################################################################################################
        
   #There are many images that are almost fully covered in dense clouds. We thus first filter and mask clouds in the image collection.
   #We delete all images with a cloud cover fraction >= max_cloud_fraction. In the remaining images we mask all pixels that are 
   #covered by clouds.
    if(mask_clouds==TRUE){   
        
      #print message   
      print("Cloud masking = TRUE")
      
      #(11) Add cloud information within the aoi_Shapefile to the image collection: 

        #Add the cloud fraction within the shapefile area to each separate image by mapping the cloud functions over the image collection
        s2_col <- s2_col$
          #Determine which pixels are clouds using the cloud_probability property
          map(compute_Clouds_cldprb)$
          #Add the fraction of cloud-covered pixels within the shapefile area as image property
          map(Add_CloudFraction)$
          #Add NULL to those images in which cloudfraction could not be calculated
          map(Add_NULL_CloudFraction)$
          #Add 'seconds since 1st of january' as a new band to each image to make sure that each pixel in the composite image can be traced back to the original image.
          map(add_Time)
        
        #Check if CloudFraction property has been added to each image
        #s2_col$first()$propertyNames()$getInfo()
        
        # #Visual check of cloud mask (for debugging)
        # image <- s2_col$filterDate(paste0(year_ID, "-05-31"), paste0(year_ID, "-09-14"))$first()$
        #   clipToCollection(aoi_Shapefile)$
        #   select("clouds_unadjusted", "clouds", "clouds_inv", "clouds_prob", "B4", "B3", "B2")      
        # Map$setCenter(coordinates_point$getInfo()$coordinates[1], coordinates_point$getInfo()$coordinates[2], 10)
        # Map$addLayer(image,list(bands=c("B4", "B3", "B2"), min=100, max=8000, gamma=c(1.9, 1.7, 1.7)), 'TRUE COLOR')+
        #   #Map$addLayer(image, list(bands='clouds_prob', min=0, max=100, opacity=1, palette=c('000000', '0dffff', '0524ff', 'ffffff')), 'clouds_prob')+
        #   Map$addLayer(image, list(bands='clouds_unadjusted', min=0, max=1, opacity=0.5, palette=c('000000', 'orange')), 'clouds unadjusted')+
        #   Map$addLayer(image, list(bands='clouds', min=0, max=1, opacity=0.5, palette=c('000000', 'orange')), 'clouds buffered')
        # #Map$addLayer(image, list(bands='clouds_inv', min=0, max=1, opacity=1), 'clouds_inv')
        
        #We have now calculated the fraction of cloud covered pixels within our shapefile area for each image in the image collection.
        #However, Sentinel-2 satellite data is recorded as different tiles, these tiles are generally much bigger than a single study site.
        #Each Sentinel-2 tile also contains a native measure of cloud cover. This is the property 'CLOUDY_PIXEL_PERCENTAGE'. This property
        #thus refers to the percentage of cloud cover within the whole tile. We have now calculated a more appropriate measure for our
        #shapefile area only called 'CloudFraction'. Below we compare both measures, as we expect they should be correlated. Differences
        #between both measures can for example arise when the shapefile area was covered by clouds while most of the larger region (tile)
        #was not or vice versa.
        
        # #Extract cloud fraction of all images in image collection for the aoi_Shapefile
        # cloud_aoi <- unlist(s2_col$aggregate_array('CloudFraction')$getInfo())
        # 
        # #Extract cloud fraction of all images in image collection for the complete tile to which our aoi_Shapefile belongs
        # cloud_tile <- unlist(s2_col$aggregate_array('CLOUDY_PIXEL_PERCENTAGE')$getInfo())
        # 
        # #Replace -9999 values by NA
        # cloud_aoi[cloud_aoi < -9000] <- NA
        # cloud_tile[cloud_tile < -9000] <- NA
        # 
        # #Assess relationship between both measures:
        # ggplot()+
        #   geom_point(aes(x=cloud_aoi, y=cloud_tile))+
        #   stat_smooth(method='lm', formula=y~x)+
        #   theme_tom()
        #   #Thus, this indicates that our local cloud measure does not necessarily correspond to that as measured on the whole tile as
        #   #explained above.
      
       #(12): Exclude all images from the image collection that have a CloudFraction value > max_cloud_fraction, and mask all cloud pixels within the remaining images
    
        #In the analysis for the whole study area we exclude images with more than 20% snowcover within our shapefile 
        #However, now we're interested in individual pixels. So even when 90% of the shapefile area is snowcovered,
        #the pixels we're interested in might be located in the 10% cloud free area. We sthus only filter images with
        #a 100% cloudcover (cloudFraction=1.0). We apply a cloud mask to mask out cloudy pixels within each image.

        #Mask cloudy pixels within each image: 
         s2_clouds_filtered <- s2_col$
            #Only remove images that are fully cloud-covered
            filter(ee$Filter$lt('CloudFraction', max_cloud_fraction))$
            #Apply cloudmask for individual pixels
            map(Add_CloudMask)
        
       }
    if(mask_clouds==FALSE){
          
       #print message   
       print("Cloud masking = FALSE")
          
       #Add Date and Time characteristics to each image
       s2_clouds_filtered <- s2_col$
          #Add 'seconds since 1st of january' as a new band to each image to make sure that each pixel in the composite image can be traced back to the original image.
          map(add_Time)
   
       }
             
##################################################################################################################################

#V: Mask permanent waterbodies (ponds, lakes, rivers, sea) within the image collection

##################################################################################################################################

   #Mask permanent waterbodies if mask_water==TRUE
    if(mask_water==TRUE){
           
       #(A): print message   
       print("Water masking = TRUE")
           
       #(B): Method 1: Extract permanent waterbodies from the ESA WorldCover map
       if(mask_water_type=="water_mask_ESA" | mask_water_type=="both"){
             
         #print message   
         print("Water masking type = ESA")
             
         #Load auxilliary function
         compute_Water_ESA=compute_Water_ESA #sourced
             
         #Extract pixels corresponding to permanent waterbodies 
         water_mask1 <- compute_Water_ESA(ESA_index=80)
             
         }
           
       #(C): Method 2: Manual detection of permanent waterbodies based on NDWI, NDSI and NIR bands
       if(mask_water_type=="water_mask_Manual" | mask_water_type=="both"){
             
         #print message   
         print("Water masking type = Manual")
             
         #Load auxilliary function
         compute_Water_Manual=compute_Water_Manual #sourced
             
         #Extract pixels corresponding to permanent waterbodies
         water_mask2 <- compute_Water_Manual(img_col=s2_col, start_date_NDWI=start_date_NDWI, end_date_NDWI=end_date_NDWI, NDWI_threshold=NDWI_threshold,
                                             NDSI_threshold=NDSI_threshold_vector[1], NIR_threshold=NIR_threshold)
             
             
         }
           
       #(D): Apply the water masking function to each image in the collection:
           
         #Define the final water_mask:
         if(mask_water_type=="water_mask_ESA"){water_mask <- water_mask1}
         if(mask_water_type=="water_mask_Manual"){water_mask <- water_mask2}
         if(mask_water_type=="both"){water_mask <- water_mask1$Or(water_mask2)}     
           
         # #Plot permanent waterbodies(for debugging)
         # Map$setCenter(coordinates_point$getInfo()$coordinates[1], coordinates_point$getInfo()$coordinates[2], 10)
         # Map$addLayer(s2_col$filterDate(paste0(year_ID, "-08-04"), paste0(year_ID, "-09-18"))$first(), list(bands=c("B4", "B3", "B2"), min=100, max=8000, gamma=c(1.9, 1.7, 1.7)), 'TRUE COLOR')+
         # Map$addLayer(water_mask,list(min=0, max = 1, palette = c('ffffff', '000000')), 'Water_mask')
           
         #Load auxilliary function to mask water-pixels
         Add_WaterMask=Add_WaterMask #sourced
           
         #Apply the final watermask:
         s2_clouds_filtered <- s2_clouds_filtered$map(Add_WaterMask)
           
       }
    if(mask_water==FALSE){
           
      #print message   
      print("Water masking = FALSE")
           
     }

##################################################################################################################################

#VI: Create a composite image for each day by mosaicking all images from the that day

##################################################################################################################################

  #There might be days for which multiple satellite photos are available. We thus need to make a decision on what to do with overlapping
  #values per pixel. We use a qualityMosaic that can select the overlapping pixel with the least cloudcover by tapping into the band
  #"cloud_inv". If two pixels have equal cloudcover, then the pixelvalue from this first image is chosen.

  #Note that cloud/water masking is taken into account when making the composite image. The reducer function "qualitymosaic" will  for every pixel only use
  #the unmasked pixels to calculate the output value (i.e. if for a certain pixel there are 4 overlapping images available for the same date and one
  #of them is covered by clouds (and thus masked), then only the remaining cloud-free (and unmasked) pixels are used to calculate the band values
  #at that pixel). See, https://gis.stackexchange.com/questions/337666/applying-google-earth-engine-cloud-mask-on-each-individual-image-before-mosaicin,
  #or: https://gis.stackexchange.com/questions/371356/combining-single-cloud-masked-images-using-google-earth-engine, or
  #https://open-mrv.readthedocs.io/en/latest/image_composite_Web.html.

   #F.1: Make a composite image for a single date (doy) and calculate fraction of snowcover - USED FOR DEBUGGING ONLY
    {
    # #Pick a date and check whether there are multiple tiles spanning the aoi
    # doy=165
    # s2_img <- s2_clouds_filtered$filterMetadata('year', 'equals', as.numeric(year_ID))$filterMetadata('doy', 'equals', doy)
    # s2_img <- s2_img$select(c('B4', 'B3', 'B2', 'NDSI', 'clouds_inv'))
    # 
    # # #Number of images in s2_clouds_filtered
    # # s2_img$size()$getInfo()
    # # s2_img$first()$bandNames()$getInfo()
    # # s2_img$first()$propertyNames()$getInfo()
    # 
    # # #Inspect whether there is any overlap between the tiles for this day:
    # # Map$setCenter(coordinates_point$getInfo()$coordinates[1], coordinates_point$getInfo()$coordinates[2], 10)
    # # Map$addLayers(s2_img, list(bands=c("B4", "B3", "B2"), min=100, max=12000, gamma=c(1.9, 1.7, 1.7)), 'S2_Original')+
    # # #Map$addLayers(s2_img, list(bands=c("seconds"), min=15053255, max=15053280), opacity=0.75, 'image_ID')
    # # Map$addLayer(aoi)
    # 
    # #Create a composite image for this day
    # s2_composite <- ee$Image(s2_img$
    #                            #Create a quality mosaic based on cloud_inv
    #                            qualityMosaic('clouds_inv')$
    #                            #Copy timestamp properties of first image to mosaic image
    #                            copyProperties(s2_img$first(), c('system:id', 'system:time_start')))#$reproject(crs=crs, crsTransform=NULL, scale=resolution)
    # s2_composite <- s2_composite$clip(aoi)
    # #s2_composite$select('NDSI')$projection()$getInfo()
    # #Note that there is no need to reproject as this is done automatically when plotting the data, or when extracting other output.
    # 
    # #s2_composite$bandNames()$getInfo()
    # #s2_composite$propertyNames()$getInfo()
    # #s2_composite$select('NDSI')$projection()$getInfo()
    # 
    # # #Visual checks
    # #  #Inspect the RGB band of the Mosaicked image
    # #  Map$setCenter(coordinates_point$getInfo()$coordinates[1], coordinates_point$getInfo()$coordinates[2], 10)
    # #  Map$addLayer(s2_composite, list(bands=c("B4", "B3", "B2"), min=100, max=12000, gamma=c(1.9, 1.7, 1.7)), 'S2_Median')
    # #
    # #  #Inspect the "clouds" band of the Mosaicked image
    # #  Map$setCenter(coordinates_point$getInfo()$coordinates[1], coordinates_point$getInfo()$coordinates[2], 10)
    # #  Map$addLayer(s2_composite, list(bands=c("clouds_inv"), palette=c("red", "green"), min=0, max=1), opacity=1, 'Clouds_median')
    # #
    # #  #Inspect to which original image each pixel in the Mosaicked image belongs (using the "seconds" band)
    # #  Map$addLayer(s2_composite, list(bands=c("seconds"), min=15053255, max=15053280), opacity=1, 'image_ID')
    # 
    # #Calculate fraction of snowcover for this composite image:
    # 
    #   #Select the NDSI band
    #   ndsi <- s2_composite$select('NDSI')
    # 
    #   #Create a binary layer using logical operations.
    #   snow <- ndsi$gt(NDSI_threshold_vector[1])$rename('SNOW')
    # 
    #   #Return the binary snow parameter
    #   s2_composite <- s2_composite$addBands(snow)
    # 
    #   # #Mask and display the binary layer (for debugging).
    #   #  Map$setCenter(coordinates_point$getInfo()$coordinates[1], coordinates_point$getInfo()$coordinates[2], 10)
    #   #  Map$addLayer(s2_composite,list(bands=c("B4", "B3", "B2"), min=0, max=12000, gamma=c(1.9, 1.7, 1.7)), 'TRUE COLOR')+
    #   #  Map$addLayer(s2_composite$select('SNOW'))
    # 
    #   #Calculate the fraction of pixels within aoi that are marked as snow
    #   SnowFraction = snow$reduceRegion(
    #     reducer = ee$Reducer$mean(),
    #     geometry = aoi_Shapefile,
    #     scale = resolution
    #     )$get('SNOW')
    # 
    #   # #Print snowfraction
    #   # SnowFraction$getInfo()
    # 
    #   #Add SnowFraction as an image property to each image
    #   s2_composite <- s2_composite$set('SnowFraction', SnowFraction)
    # 
    #   #Set the SnowFraction value to a no data value of -9999 if equal to NULL
    #   SnowFraction <- ee$List(list(s2_composite$get('SnowFraction'), -9999))$reduce(ee$Reducer$firstNonNull())
    # 
    #   #Return the updated SnowFraction property
    #   s2_composite <- s2_composite$set("SnowFraction", SnowFraction)
    # 
    #   #s2_composite$propertyNames()$getInfo()
    }

   #F.2: Create a median composite image for each day of year (doy) per year
   #Create a list of unique doy per year, loop through this list and add all images belonging to each doy to this list, group
   #images and create a composite image by reducing each pixel value using a qualitymosaic based on cloudcover.
   if(create_composite==TRUE){

    #print message
     print("Composite image = TRUE")

    #Define a list of unique years and doy from the image collection.
     years <- ee$List(s2_clouds_filtered$aggregate_array('year'))$distinct()$sort()
     days <- ee$List(s2_clouds_filtered$aggregate_array('doy'))$distinct()$sort()

    #Map over a the list of each distinct doy, and year to build a list of image mosaics
     s2_colList <- years$map(ee_utils_pyfunc(function(year){

      return(days$map(ee_utils_pyfunc(function(doy){

        #Select all images for current doy and year combination
        selection <- s2_clouds_filtered$
          #Filter image collection by year
          filterMetadata('year', 'equals', year)$
          #Filter image collection by doy
          filterMetadata('doy', 'equals', doy)$
          #Select required bands only
          select(c('B4', 'B3', 'B2', 'NDSI', 'NDVI', 'NDMI', 'NDWI', 'clouds', 'clouds_inv'))
          #select(c('NDSI', 'clouds_inv'))

        #Create a composite image for the current doy and year combination
        selection_mosaic <- ee$Image(selection$
                                       #Create Quality mosaic
                                       qualityMosaic('clouds_inv')$ #median()$ #mosaic() #reduce(ee$Reducer$qualityMosaic('clouds_inv'))
                                       #Reproject mosaicked image
                                       #reproject(crs=crs, crsTransform=NULL, scale=resolution)$
                                       #Copy timestamp properties of first image to mosaic image
                                       copyProperties(selection$first(), c('system:id', 'system:time_start')))

        #Return composite median image (which is stored in list)
        return(selection_mosaic$
                 #Set doy as an image property
                 set('doy', doy)$
                 #Set year as an image property
                 set('year', year)
               )})))}))$flatten()

    #Convert the image List to an ImageCollection.
     s2_col_composite <- ee$ImageCollection$fromImages(s2_colList)

    #Inspect the new composite image collection
     #s2_clouds_filtered$size()$getInfo() #before compositing
     #s2_col_composite$size()$getInfo() #after compositing
     #s2_col_composite$first()$bandNames()$getInfo()
     #s2_col_composite$first()$propertyNames()$getInfo()
     #s2_col_composite$aggregate_stats("system:time_start")$getInfo()
     #s2_col_composite$first()$select('NDSI')$projection()$getInfo()

     #Note that the default projection of a composite image is "EPSG:4326" with a scale of 1 degree. This is NOT a
     #problem as we specify  the required output projection at all reduceRegions and export steps below. See:
     #https://developers.google.com/earth-engine/guides/projections#the-default-projection
     #Alternatively, we can specify the reproject function several lines above (which is now commented)

   }
   if(create_composite==FALSE){

     #print message
     print("Composite image = FALSE")

     #set s2_col_composite to s2_clouds_filtered (i.e. to unmosaicked image)
     s2_col_composite <- s2_clouds_filtered

   }
      
##################################################################################################################################

#VII: Count the total number of unmasked pixels per doy within aoi_Shapefile

##################################################################################################################################

  #Count the total number of unmasked pixels and the total number of pixels per doy within aoi_Shapefile
   if(pixel_counts==TRUE){
      
     #(A): Add pixel counts within the area of interest to each separate image by mapping the pixel count functions over the image collection
      s2_col_composite <- s2_col_composite$
        #Count number of unmasked pixels within aoi_Shapefile
        map(AddPixelCount)$
        #Add NULL to those images in which pixel count could not be calculated
        map(AddNULLPixelCount)

     #(B): Extract pixel_counts of all images in image collection for aoi_Shapefile

       #create a current timestamp to prevent identical names on Google Drive
        current_timestamp0 <- gsub('\\.', '', format(Sys.time(), "%y%m%d%H%M%OS2"))

       #We use ee_table_to_drive() to prevent memory limits
        task_vector0 <- ee_table_to_drive(
          collection = s2_col_composite,
          description = paste0(current_timestamp0, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Data_Pixel_Counts_polygon"),
          fileFormat = "CSV",
          selectors = c('doy', 'unmasked', 'total')
          )

       #Monitor the task
        task_vector0$start()
        print("Count the number of unmasked pixels within the shapefile per doy:")
        ee_monitoring(task_vector0, task_time=60, max_attempts=1000000)

       #Export results to local folder
        exported_stats <- ee_drive_to_local(task = task_vector0, dsn=paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Data_Pixel_Counts_polygon"))
        df_pixelcount <- read.csv(exported_stats)
        unlink(exported_stats)

     #(C): Replace -9999 values by NA (for debugging)
      df_pixelcount$unmasked[df_pixelcount$unmasked < -9000] <- NA
      df_pixelcount$total[df_pixelcount$total < -9000] <- NA
      df_pixelcount$doy[df_pixelcount$doy < -9000] <- NA

     #(D): Add missing dates with 0 unmasked pixels to the dataframe
      df_doy_missing <- data.frame(doy=seq(start_date_doy, end_date_doy)[!(seq(start_date_doy, end_date_doy) %in% df_pixelcount$doy)],
                                   unmasked=0,
                                   total=max(df_pixelcount$total))
      df_pixelcount <- rbind(df_pixelcount, df_doy_missing)
      df_pixelcount <- df_pixelcount[order(df_pixelcount$doy),]
      write.csv(df_pixelcount, file=paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Data_Pixel_Counts_polygon.csv"), quote=FALSE, row.names=FALSE)

     #(E): Create barplot with the pixel counts per day of year within aoi_Shapefile
      df_pixelcount$masked <- df_pixelcount$total - df_pixelcount$unmasked
      df_pixelcount <- df_pixelcount[,-which(colnames(df_pixelcount) %in% "total")]
      df_pixelcount <- tidyr::pivot_longer(df_pixelcount, cols=c(unmasked, masked), names_to = "pixels")
      p_pixelcounts <- ggplot()+
       geom_bar(data=df_pixelcount, aes(x=doy, y=value, fill=pixels), stat="identity",colour="black", position="stack", width=1)+
       #geom_rect(aes(xmin=min(df_pixelcount$doy)-1, xmax=max(df_pixelcount$doy)+1, ymin=0, ymax=max(df_pixelcount$value)), alpha=0.5, fill="white")+
       scale_fill_manual(values = c("black", "#FFA52C"))+
       xlab("Time (day of year)")+
       ylab("Pixel count")+
       theme_classic()

     #(F): Save barplot
      pdf(paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Plot_Pixel_Counts_polygon.pdf"), width=12, height=8)
      print(p_pixelcounts)
      dev.off()
      
   }

##################################################################################################################################

#VIII: Create timelapse videos of the RGB bands, and the NDSI-, NDVI-, NDMI- and NDMI-bands for aoi_Polygons (Cloud and Water masked)

##################################################################################################################################

  #Note that these GIFs generally result in memory errors with areas larger than 100km2. Adjust dimensions to get 
  #higher resolution GIFS.
  
  #(14): Create timeseries GIF animations
   if(gif_output==TRUE){
  
    #(a): GIF of RGB-images
    f_img_col_to_gif(img_col=s2_col_composite,  #image collection
                     RGB_bands=c("B4", "B3", "B2"), #names of RGB-bands in image collection
                     shapefile=aoi_Shapefile, #shapefile of area of interest
                     centroid_buffer_m=0, #buffer zone surrounding centroid of shapefile in meters (to enlarge area for GIF)
                     gif_dimensions=gif_max_pixels, #Maximum dimensions of GIF (pixels)
                     gif_fps=5, #Frames per second of GIF
                     gif_min=0, #Value to map to 0 (to improve contrast). A good rule of thumb is to set min to values that represent the 2nd percentile of the data
                     gif_max=12000, #Value to map to 255 (to improve contrast). A good rule of thumb is to set max to values that represent the 98th percentile of the data
                     gif_gamma=c(1.9, 1.7, 1.7), #Gamma correction factors (one for each band)
                     gif_crs='EPSG:3857', #CRS project of the output
                     gif_text_position="northwest", #Location of text (datetime string)
                     gif_text_position_adjustment="+0+0", #Small scale adjustment of text in meters
                     gif_text_size=14, #Size of text
                     gif_text_col="#FFFFFF",
                     output_fldr=paste0(root_fldr, "/Output/S2/08_Shapefile_Points_Snowmelt/"),
                     file_name="GIF_RGB_masked")
    
    #(b): GIF of NDSI-images
    
      #Transform NDSI-band to RGB bands
      s2_col_composite_ndsi <- s2_col_composite$map(ee_utils_pyfunc(function(image){
        f_band_to_RGB(img=image, band='NDSI', min_value=-1, max_value=1.5, palette=c('#000000', '#0dffff', '#0524ff', '#ffffff'))}))
      
      #Create gif of the transformed NDSI-band
      f_img_col_to_gif(img_col=s2_col_composite_ndsi,  #image collection
                       RGB_bands=c("vis-red", "vis-green", "vis-blue"), #names of RGB-bands in image collection
                       shapefile=aoi_Shapefile, #shapefile of area of interest
                       centroid_buffer_m=0, #buffer zone surrounding centroid of shapefile in meters (to enlarge area for GIF)
                       gif_dimensions=gif_max_pixels, #Maximum dimensions of GIF (pixels)
                       gif_fps=5, #Frames per second of GIF
                       gif_min=0, #Value to map to 0 (to improve contrast). A good rule of thumb is to set min to values that represent the 2nd percentile of the data
                       gif_max=255, #Value to map to 255 (to improve contrast). A good rule of thumb is to set max to values that represent the 98th percentile of the data
                       gif_gamma=c(1, 1, 1), #Gamma correction factors (one for each band)
                       gif_crs='EPSG:3857', #CRS project of the output
                       gif_text_position="northwest", #Location of text (datetime string)
                       gif_text_position_adjustment="+0+0", #Small scale adjustment of text in meters
                       gif_text_size=14, #Size of text
                       gif_text_col="#FFFFFF",
                       output_fldr=paste0(root_fldr, "/Output/S2/08_Shapefile_Points_Snowmelt/"),
                       file_name="GIF_NDSI_masked")
    
    # #(c): GIF of NDVI-images
    #    
    #    #Transform NDVI-band to RGB bands
    #    s2_col_composite_ndvi <- s2_col_composite$map(ee_utils_pyfunc(function(image){
    #      f_band_to_RGB(img=image, band='NDVI', min_value=-0.25, max_value=1, palette=c("#cccccc", "#f46d43", "#fdae61", "#fee08b", "#d9ef8b", "#a6d96a", "#66bd63", "#1a9850"))}))
    #    
    #    #Create gif of the transformed NDVI-band
    #    f_img_col_to_gif(img_col=s2_col_composite_ndvi,  #image collection
    #                     RGB_bands=c("vis-red", "vis-green", "vis-blue"), #names of RGB-bands in image collection
    #                     shapefile=aoi_Shapefile, #shapefile of area of interest
    #                     centroid_buffer_m=0, #buffer zone surrounding centroid of shapefile in meters (to enlarge area for GIF)
    #                     gif_dimensions=gif_max_pixels, #Maximum dimensions of GIF (pixels)
    #                     gif_fps=5, #Frames per second of GIF
    #                     gif_min=0, #Value to map to 0 (to improve contrast). A good rule of thumb is to set min to values that represent the 2nd percentile of the data
    #                     gif_max=255, #Value to map to 255 (to improve contrast). A good rule of thumb is to set max to values that represent the 98th percentile of the data
    #                     gif_gamma=c(1, 1, 1), #Gamma correction factors (one for each band)
    #                     gif_crs='EPSG:3857', #CRS project of the output
    #                     gif_text_position="northwest", #Location of text (datetime string)
    #                     gif_text_position_adjustment="+0+0", #Small scale adjustment of text in meters
    #                     gif_text_size=14, #Size of text
    #                     gif_text_col="#FFFFFF",
    #                     output_fldr=paste0(root_fldr, "/Output/S2/08_Shapefile_Points_Snowmelt/"),
    #                     file_name="GIF_NDVI_masked")
    # 
    # #(d): GIF of NDMI-images
    #    
    #    #Transform NDMI-band to RGB bands
    #    s2_col_composite_ndmi <- s2_col_composite$map(ee_utils_pyfunc(function(image){
    #      f_band_to_RGB(img=image, band='NDMI', min_value=-0.75, max_value=1, palette=c("#d73027", "#f46d43", "#fdae61", "#fee08b", "#d9ef8b", "#a6d96a", "#66bd63", "#1a9850", "#6ad99e", "#387ad9", "#003dd6"))}))
    #    
    #    #Create gif of the transformed NDMI-band
    #    f_img_col_to_gif(img_col=s2_col_composite_ndmi,  #image collection
    #                     RGB_bands=c("vis-red", "vis-green", "vis-blue"), #names of RGB-bands in image collection
    #                     shapefile=aoi_Shapefile, #shapefile of area of interest
    #                     centroid_buffer_m=0, #buffer zone surrounding centroid of shapefile in meters (to enlarge area for GIF)
    #                     gif_dimensions=gif_max_pixels, #Maximum dimensions of GIF (pixels)
    #                     gif_fps=5, #Frames per second of GIF
    #                     gif_min=0, #Value to map to 0 (to improve contrast). A good rule of thumb is to set min to values that represent the 2nd percentile of the data
    #                     gif_max=255, #Value to map to 255 (to improve contrast). A good rule of thumb is to set max to values that represent the 98th percentile of the data
    #                     gif_gamma=c(1, 1, 1), #Gamma correction factors (one for each band)
    #                     gif_crs='EPSG:3857', #CRS project of the output
    #                     gif_text_position="northwest", #Location of text (datetime string)
    #                     gif_text_position_adjustment="+0+0", #Small scale adjustment of text in meters
    #                     gif_text_size=14, #Size of text
    #                     gif_text_col="#FFFFFF",
    #                     output_fldr=paste0(root_fldr, "/Output/S2/08_Shapefile_Points_Snowmelt/"),
    #                     file_name="GIF_NDMI_masked")
    # 
    # #(e): GIF of NDWI-images
    #    
    #    #Transform NDWI-band to RGB bands
    #    s2_col_composite_ndwi <- s2_col_composite$map(ee_utils_pyfunc(function(image){
    #      f_band_to_RGB(img=image, band='NDWI', min_value=-0.5, max_value=1, palette=c('#000000', '#0dffff', '#0524ff', '#ffffff'))}))
    #    
    #    #Create gif of the transformed NDWI-band
    #    f_img_col_to_gif(img_col=s2_col_composite_ndwi,  #image collection
    #                     RGB_bands=c("vis-red", "vis-green", "vis-blue"), #names of RGB-bands in image collection
    #                     shapefile=aoi_Shapefile, #shapefile of area of interest
    #                     centroid_buffer_m=0, #buffer zone surrounding centroid of shapefile in meters (to enlarge area for GIF)
    #                     gif_dimensions=gif_max_pixels, #Maximum dimensions of GIF (pixels)
    #                     gif_fps=5, #Frames per second of GIF
    #                     gif_min=0, #Value to map to 0 (to improve contrast). A good rule of thumb is to set min to values that represent the 2nd percentile of the data
    #                     gif_max=255, #Value to map to 255 (to improve contrast). A good rule of thumb is to set max to values that represent the 98th percentile of the data
    #                     gif_gamma=c(1, 1, 1), #Gamma correction factors (one for each band)
    #                     gif_crs='EPSG:3857', #CRS project of the output
    #                     gif_text_position="northwest", #Location of text (datetime string)
    #                     gif_text_position_adjustment="+0+0", #Small scale adjustment of text in meters
    #                     gif_text_size=14, #Size of text
    #                     gif_text_col="#FFFFFF",
    #                     output_fldr=paste0(root_fldr, "/Output/S2/08_Shapefile_Points_Snowmelt/"),
    #                     file_name="GIF_NDWI_masked")
    
}
      
##################################################################################################################################
      
#IX: Iterate through all locations of interest and extract (I) mean bandvalues, (II) the fraction of snowcover within each point's bufferzone over time 
      
##################################################################################################################################       
     
 #Method 'avg_NDSI'       
  if("avg_NDSI" %in% method){       
         
      #Extract Sentinel-2 band values for points of interest (with buffer) for all images in the image collection (output a feature collection)  
         
       #Create an iteration function that we will use to iterate through all images of the image collection. For each image, the value
       #of certain bands is extracted at each feature (i.e. location) within the feature collection Locations. The resulting
       #band values (and the datetime of the image) are added as a property to each feature (i.e. location). This results
       #in an updated feature collection specific for the current image. This feature collection is then appended to a list
       #of feature collections from previous image iterations. After iterating through n-images, the result is thus a feature
       #collection list of length n * the number of features within the feature collection.
           
       #Thus, at each iteration we extract band values of interest of all locations for the current image, store this as a  
       #feature collection and add this to an expanding list of feature collections from previous iterations.
           
       #Store default Sentinel-2 image projection 
        S2Projection <- s2_col_composite$first()$select("NDSI")$projection()
        #S2Projection$getInfo() 
       
       #Create an empty FeatureCollection list. This list is used as input for the first iteration of the iteration function below.
        FC_initial_meanbandvalues <- ee$FeatureCollection(ee$List(list()))
           
       #Specify the iteration function. This function takes two arguments. The first argument is the current element of the image collection
       #(in this case the current iteration image) and the second element takes the output value from the iteration that preceded it. The latter
       #is not possible for the first iteration, that's why an initial object (empty feature collection) to start the iteration with 
       #has been defined.
        Extract_BandValuesAtPoins = Extract_BandValuesAtPoins #sourced
           
       #Iterate over the ImageCollection and select appropriate band values
        FC_merged_meanbandvalues <- ee$FeatureCollection(s2_col_composite$select("NDSI", "NDVI", "NDMI")$iterate(Extract_BandValuesAtPoins, FC_initial_meanbandvalues))
         
      #Transform feature collection with Band values to a dataframe     
           
        #Export the feature collection as a .csv table
        #We export the data instead of using aggregate_array() as the latter might fail due to computation timeouts.
        
        #create a current timestamp to prevent identical names on Google Drive
        current_timestamp1 <- gsub('\\.', '', format(Sys.time(), "%y%m%d%H%M%OS2"))
        
        #Setup task
        task_vector1 <- ee_table_to_drive(
          collection= FC_merged_meanbandvalues,
          description = paste0(current_timestamp1, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Data_MeanBandvalues"),
          folder="RGEE_tmp",
          fileFormat="CSV",
          selectors=c('NDSI', 'NDVI', 'NDMI', 'Date', 'LocationID')
        )
        
        #Run and monitor task
        task_vector1$start()
        ee_monitoring(task_vector1, task_time=60, max_attempts=1000000) #250s at 100m
        
        #Import results
        exported_stats <- ee_drive_to_local(task=task_vector1, dsn=paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Data_MeanBandvalues"))
        Locations_Data_BandValues <- read.csv(exported_stats)
        unlink(exported_stats)
        
        #Add day of year
        Locations_Data_BandValues$Date <- as.POSIXct(Locations_Data_BandValues$Date, format="%Y-%m-%d %H:%M:%S")
        Locations_Data_BandValues$doy <- as.numeric(strftime(Locations_Data_BandValues$Date, format = "%j"))
        
        #Remove NAs in the NDSI, NDVI and NDMI variables
        Locations_Data_BandValues$NDSI[Locations_Data_BandValues$NDSI < -9000] <- NA
        Locations_Data_BandValues$NDVI[Locations_Data_BandValues$NDVI < -9000] <- NA
        Locations_Data_BandValues$NDMI[Locations_Data_BandValues$NDMI < -9000] <- NA
        
        #Sort dataframe by LocationID and doy
        index <- with(Locations_Data_BandValues, order(LocationID, doy))
        Locations_Data_BandValues <- Locations_Data_BandValues[index,]
        Locations_Data_BandValues <- Locations_Data_BandValues[,c("LocationID", "Date", "doy", "NDSI", "NDVI", "NDMI")]
        
        #Add coordinates to each location
        Locations_Data_BandValues <- left_join(Locations_Data_BandValues, Locations_sf[,c("LON_x", "LAT_y", "LocationID")], by=c("LocationID"))
        
        #Save dataframe
        write.csv(Locations_Data_BandValues, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Data_MeanBandvalues.csv"), row.names = FALSE)
        
    }
      
 #Method 'snowfraction'
  if("snowfraction" %in% method){

    #Create an empty dataframe for storing the Snowfraction results per location (method="SnowFraction")
    df_Locations_SnowFraction <- data.frame(LocationID=character(),
                                            Date=character(),
                                            doy=numeric(),
                                            SnowFraction=numeric(),
                                            NDSI_threshold=character(),
                                            LON_x=numeric(),
                                            LAT_y=numeric())

    #Run the analysis for each level of NDSI_threshold_vector
    for(NDSI_threshold in NDSI_threshold_vector){

      #for debugging
      #NDSI_threshold <- NDSI_threshold_vector[1]

      #Print message
      print(paste0("  -Start analysis for NDSI_threshold = ", NDSI_threshold))

      #Store NDSI_threshold as a character (used for naming of outputs)
      NDSI_threshold_char <- gsub("\\.", "_", as.character(NDSI_threshold))

     #Add a binary snow cover band to each image and calculate the fraction of snow covered pixels within each point's buffer zone

      #Map Snow computation functions over the mosaicked image collection
      s2_col_composite_snow <- s2_col_composite$
        #Determine which pixels are snow-covered (NDSI > NDSI threshold)
        map(computeSnow)

     #Create an iteration function that we will use to iterate through all images of the image collection. For each image, the
     #average value of the "SNOW" band is extracted at each feature (i.e. location) within the feature collection Locations.
     #The resulting fractino of snowcover and the datetime of the image are added as a property to each feature (i.e. location).
     #This results in an updated feature collection specific for the current image. This feature collection is then appended to a list
     #of feature collections from previous image iterations. After iterating through n-images, the result is thus a feature
     #collection list of length n * the number of features within the feature collection.

     #Thus, at each iteration we extract the average value of the "SNOW" band (i.e. the fraction of snowcover as it comprises of
     #zeros and ones only) for all locations for the current image, store this as a  feature collection and add this to an expanding
     #list of feature collections from previous iterations.

     #Create an empty FeatureCollection list. This list is used as input for the first iteration of the iteration function below.
     FC_initial_snowfraction <- ee$FeatureCollection(ee$List(list()))

     #Specify the iteration function. This function takes two arguments. The first argument is the current element of the image collection
     #(in this case the current iteration image) and the second element takes the output value from the iteration that preceded it. The latter
     #is not possible for the first iteration, that's why an initial object (empty feature collection) to start the iteration with
     #has been defined.
     Extract_SnowFractionAtPoints = function(img, FC_initial_snowfraction) {

             #(A): Extract Snowfraction within the current image at each point in the Locations feature collection. We add this
             #fraction as a property to each feature (i.e. location). The resulting output is a feature collection for the current
             #img where the snowfraction is added as property for each location.
             FC_image <- img$reduceRegions(collection = Locations,
                                           reducer = ee$Reducer$mean()$setOutputs(list("SnowFraction")),
                                           scale = resolution, #implicitly specified in crs below
                                           crs=crs, #or, crs=S2Projection
                                           crsTransform=NULL)

             #(B): reduceRegions does not return any SnowFraction value when the point is masked by clouds or does not fall within
             #the boundaries of the image. Therefore we manually have to add an empty SnowFraction property to those features. We
             #therefore set the SnowFraction value to a no data value of -9999 for all features where the SnowFraction value is NULL.
             FC_image <- FC_image$map(function(feature){
               snowfraction <- ee$List(list(feature$get('SnowFraction'), -9999))$reduce(ee$Reducer$firstNonNull())
               return(feature$set("SnowFraction", snowfraction))
               })

             #(C): Add date and day of year of current image as a property to each feature within the feature collection FC_image
             date <- img$date()$format("YYYY-MM-dd hh:mm:ss")
             doy <- ee$Date(img$get('system:time_start'))$getRelative('day', 'year')
             FC_image <- FC_image$map(function(feature){return(feature$set("Date", date)$set("doy", doy))})

             #(D): Merge the feature collection of the current image (FC_image) onto the feature collection FC_initial_snowfraction
             return (ee$FeatureCollection(FC_initial_snowfraction)$merge(FC_image))
             #FC_initial_snowfraction is thus updated at each iteration. This corresponds to base R code:
             #FC_initial_snowfraction <- merge(FC_initial_snowfraction, FC_image)
           }

     #Iterate over the ImageCollection and select appropriate band values
     FC_merged_SnowFraction <- ee$FeatureCollection(s2_col_composite_snow$select("SNOW")$iterate(Extract_SnowFractionAtPoints, FC_initial_snowfraction))

     #Export the feature collection as a .csv table
     #We export the data instead of using aggregate_array() as the latter might fail due to computation timeouts.

       #create a current timestamp to prevent identical names on Google Drive
        current_timestamp2 <- gsub('\\.', '', format(Sys.time(), "%y%m%d%H%M%OS2"))
     
       #Setup task
        task_vector2 <- ee_table_to_drive(
          collection= FC_merged_SnowFraction,
          description = paste0(current_timestamp2, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_NDSI", NDSI_threshold_char, "_Locations_Data_SnowFraction"),
          folder="RGEE_tmp",
          fileFormat="CSV",
          selectors=c('SnowFraction', 'Date', 'LocationID')
          )

       #Run and monitor task
        task_vector2$start()
        ee_monitoring(task_vector2, task_time=60, max_attempts=1000000) #250s at 100m

       #Import results
        exported_stats <- ee_drive_to_local(task=task_vector2, dsn=paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_NDSI", NDSI_threshold_char, "_Locations_Data_SnowFraction"))
        df_Locations_SnowFraction_new <- read.csv(exported_stats)
        unlink(exported_stats)

       #Add day of year
        df_Locations_SnowFraction_new$Date <- as.POSIXct(df_Locations_SnowFraction_new$Date, format="%Y-%m-%d %H:%M:%S")
        df_Locations_SnowFraction_new$doy <- as.numeric(strftime(df_Locations_SnowFraction_new$Date, format = "%j"))

       #Remove NAs in the SnowFraction variable
        df_Locations_SnowFraction_new$SnowFraction[df_Locations_SnowFraction_new$SnowFraction < -9000] <- NA #replace -9999 by NA
        #df_Locations_SnowFraction_new <- df_Locations_SnowFraction_new[!is.na(df_Locations_SnowFraction_new$SnowFraction), ]

       #Sort dataframe by LocationID and doy
        index <- with(df_Locations_SnowFraction_new, order(LocationID, doy))
        df_Locations_SnowFraction_new <- df_Locations_SnowFraction_new[index,]
        df_Locations_SnowFraction_new <- df_Locations_SnowFraction_new[,c("LocationID", "Date", "doy", "SnowFraction")]

       #Add NDSI_threshold as a new column
        df_Locations_SnowFraction_new$NDSI_threshold <- as.factor(NDSI_threshold)

       #Add coordinates to each location
        df_Locations_SnowFraction_new <- left_join(df_Locations_SnowFraction_new, Locations_sf[,c("LON_x", "LAT_y", "LocationID")], by=c("LocationID"))

       #Add dataframe for current location to dataframe from previous iterations:
        df_Locations_SnowFraction <- rbind(df_Locations_SnowFraction, df_Locations_SnowFraction_new)

        # #For debugging
        #  ggplot() + geom_point(data=df_Locations_SnowFraction, aes(x=doy, y=SnowFraction, col=NDSI_threshold)) + facet_wrap(~LocationID) + theme_classic()

       }

    #Save dataframe
    write.csv(df_Locations_SnowFraction, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Data_SnowFraction.csv"), row.names = FALSE)
    #df_Locations_SnowFraction <- read.csv(file=paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Data_SnowFraction.csv"), header=T)

    }
      
##################################################################################################################################

#X: Fit a Generalized Additive Model (GAM) through the Mean Bandvalue and Snowfraction data for each location

##################################################################################################################################

 #Method "avg_NDSI"
  if("avg_NDSI" %in% method){

      #SNOW - Fit a Generalized Additive Model (GAM) through the NDSI data for each Location
         {
         #We fit a Generalized Additive Model (GAM) through the data for each point location. We do this using a sequential
         #outlier-filtering process. We first fit a GAM through the data, calculating model predictions and residuals. We then
         #exclude all rows from the dataframe where the residual >= (0.25 * the range of the data). We then re-fit a GAM to
         #this reduced dataset, make predictions and calculate residuals. We then exclude all rows from the reduced dataframe
         #where the residual >= (0.1 * the range of the data). This gives us a final dataframe through which we fit a third
         #GAM and store its model predictions, given the dataset in which outliers were thus removed through two subsequent
         #steps. This whole process is executed using the function f_gam_SeqRemOutliers. Note that a sequential step is
         #required because initially some datapoints might falsely be assigned a large residual because of one extreme
         #outlier. After removal of this extreme outlier and refitting of the GAM, it can be better assessed which data
         #points truly have a large residual and can thus be assigned as 'true' outliers.
            f_gam_SeqRemOutliers <- f_gam_SeqRemOutliers #sourced

            #Specify sequential outlier thresholds:
             outlier_thresh_1=outlier_thresh_1
             outlier_thresh_2=outlier_thresh_2
             outlier_removal=outlier_removal

         #(a) Specify at which NDSI value, the point is considered snow-free
             NDSI_threshold_vector = NDSI_threshold_vector

         #(b) Create an empty dataframe
             df_Locations_NDSI <- data.frame(NDSI=numeric(),
                                             Date=factor(),
                                             doy=numeric(),
                                             LocationID=factor(),
                                             outliers=logical())

             df_Locations_NDSI_predictions <- data.frame(LocationID=character(),
                                                         doy=numeric(),
                                                         NDSI_gam_predict=numeric(),
                                                         stringsAsFactors=FALSE)

         #(c) Loop through all Locations and fit a separate gam with sequential outlier removal to the location specific mean NDSI data
             for(i in unique(Locations_Data_BandValues$LocationID)){

               #For debugging
                #i=unique(Locations_Data_BandValues$LocationID)[1]

               #Select Location-specific subset of data:
                df_Location_NDSI_new <- Locations_Data_BandValues[Locations_Data_BandValues$LocationID==i & !is.na(Locations_Data_BandValues$NDSI),
                                                    c("NDSI", "Date", "doy", "LocationID")]

               #Fit a gam through the Location-specific mean_NDSI ~ doy data and employ sequential outlier removal
                df_Location_NDSI_new <- f_gam_SeqRemOutliers(data=df_Location_NDSI_new, y="NDSI", x="doy", outlier_removal=outlier_removal,
                                                    outlier_thresh_1=outlier_thresh_1, outlier_thresh_2=outlier_thresh_2,
                                                    default_k=gam_k_outlier)

               #Sort df_Location_NDSI_new by doy:
                df_Location_NDSI_new <- df_Location_NDSI_new[order(df_Location_NDSI_new$doy),]

               #Bind the Location-specific dataframe with GAM estimates to the dataframe containing all dataframes from previous iterations
                df_Locations_NDSI <- rbind(df_Locations_NDSI, df_Location_NDSI_new)

               #Create more detailed predictions (not only at the doy present in the datframe) at a 1-day interval to plot more smooth curves

                #Refit GAM through data
                 index <- which(df_Location_NDSI_new$outliers==FALSE)
                 max_k_length1 <- length(unique(df_Location_NDSI_new[index, 'doy']))
                 mod_gam <- with(df_Location_NDSI_new[index,], mgcv::gam(NDSI ~ s(doy, k=min(gam_k, length(index)-1, max_k_length1-1)), method="REML"))

                #Use gam to make predictions on a more detailed (1-day) day of year interval
                 df_Locations_NDSI_predictions_new <- data.frame(LocationID=i, doy=seq(min(df_Location_NDSI_new$doy), max(df_Location_NDSI_new$doy), 1))
                 df_Locations_NDSI_predictions_new$NDSI_gam_predict <- stats::predict(mod_gam, newdata=df_Locations_NDSI_predictions_new, type="response")
                 df_Locations_NDSI_predictions_new <- df_Locations_NDSI_predictions_new[order(df_Locations_NDSI_predictions_new$doy),]

                #Add predictions to df_Locations_NDSI_predictions dataframe:
                 df_Locations_NDSI_predictions <- rbind(df_Locations_NDSI_predictions, df_Locations_NDSI_predictions_new)

               }

             #Change column LocationID to a factor:
              df_Locations_NDSI$LocationID <- as.factor(as.character(df_Locations_NDSI$LocationID))
              df_Locations_NDSI_predictions$LocationID <- as.factor(as.character(df_Locations_NDSI_predictions$LocationID))

             #Save dataframe with GAM fits for NDSI
              #write.csv(df_Locations_NDSI, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Data_NDSI_GAM.csv"), row.names = FALSE)
              write.csv(df_Locations_NDSI_predictions, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_GAM_Predictions_NDSI.csv"), row.names = FALSE)
              
         #(d) Plot the raw NDSI datapoints and gam predictions for each Location:

             #Plot NDSI and model predictions for all locations in a single plot
              p_Locations_NDSI = ggplot()+
                geom_point(data=df_Locations_NDSI, aes(x=doy, y=NDSI, fill=LocationID, col=LocationID))+
                geom_line(data=df_Locations_NDSI_predictions, aes(x=doy, y=NDSI_gam_predict, col=LocationID)) +
                xlab(paste0("Day of year (starting at 01-01-", year_ID,")")) +
                ylab("NDSI") +
                theme_tom()
              
             #Plot NDSI and model predictions in a separate plot per location
              p_Locations_NDSI_grid = ggplot()+
                geom_point(data=df_Locations_NDSI[df_Locations_NDSI$outliers==FALSE,], aes(x=doy, y=NDSI))+
                geom_point(data=df_Locations_NDSI[df_Locations_NDSI$outliers==TRUE,], aes(x=doy, y=NDSI), col="black", pch=16, alpha=0.2)+
                geom_line(data=df_Locations_NDSI_predictions, aes(x=doy, y=NDSI_gam_predict), col="#1620de", lwd=1.25)+
                geom_vline(xintercept = 150, colour="grey", lty=2)+
                geom_hline(yintercept = NDSI_threshold_vector, colour="grey", lty=2)+
                facet_wrap(~LocationID, ncol=ceiling(length(unique(df_Locations_NDSI$LocationID))^0.5))+
                xlab(paste0("Day of year (starting at 01-01-", year_ID,")")) +
                ylab("NDSI") +
                theme_tom()

                #ggsave(plot=p_Locations_NDSI_grid, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Plot_NDSI.pdf"), width=12, height = 10)

         #(e) Calculate at which day of year the average NDSI is equal to NDSI_threshold_vector for each location using predictions from mod_gam

             #Create an empty dataframe
              df_Snowmelt_NDSI_Locations <- data.frame(LocationID=character(),
                                                       NDSI_threshold=character(),
                                                       doy=numeric(),
                                                       stringsAsFactors=FALSE)

              #Setup parallel processing
              numCores <- detectCores()
              cl <- makePSOCKcluster(numCores)
              registerDoSNOW(cl)
              
              #Extract the day of snowmelt for each location for all levels of NDSI_threshold_vector
              for(LocationID in unique(df_Locations_NDSI_predictions$LocationID)){
                
                #For debugging
                #LocationID=unique(df_Locations_NDSI_predictions$LocationID)[1] #for debugging
                
                #Use the function f_detect_threshold_date_parallel to extract the moments the GAM predictions cross the thresholds in NDSI_threshold_vector
                results <- f_detect_threshold_date_parallel(subset=1, #data subset (not relevant here, set to 1)
                                                            pixelIDs_split=LocationID, #Current location (input during loop)
                                                            df_pixel_y=df_Locations_NDSI_predictions, #dataframe containing GAM predictions
                                                            pixel_ID_column="LocationID", #Grouping column
                                                            y="NDSI_gam_predict", #response variable in GAM
                                                            x="doy", #predictor variable in GAM
                                                            pixel_gam_plots=F, #Should GAM plots be created
                                                            y_threshold=NDSI_threshold_vector) #Which threshold values for 'y' should be calculated
                
                #Store dates of snowmelt per Location
                df_Snowmelt_NDSI_Locations_new <- lapply(results, "[[", 1)
                df_Snowmelt_NDSI_Locations_new <- as.data.frame(do.call(rbind, df_Snowmelt_NDSI_Locations_new))
                colnames(df_Snowmelt_NDSI_Locations_new)[colnames(df_Snowmelt_NDSI_Locations_new)=="pixel_ID"] <- "LocationID"
                colnames(df_Snowmelt_NDSI_Locations_new)[colnames(df_Snowmelt_NDSI_Locations_new)=="x_threshold"] <- "doy"
                colnames(df_Snowmelt_NDSI_Locations_new)[colnames(df_Snowmelt_NDSI_Locations_new)=="y_threshold"] <- "NDSI_threshold"
                
                #Add snowmelt data for LocationID to general dataframe
                df_Snowmelt_NDSI_Locations <- rbind(df_Snowmelt_NDSI_Locations, df_Snowmelt_NDSI_Locations_new)
                
              }
              
              #Turn parallel processing off and run sequentially again after this point
              stopCluster(cl)
              registerDoSEQ()
              
              #Add coordinates to each location
              df_Snowmelt_NDSI_Locations <- left_join(df_Snowmelt_NDSI_Locations, Locations_sf, by=c("LocationID"))  
              
              #Save dates of snowmelt per Location as a .csv file
              write.csv(df_Snowmelt_NDSI_Locations, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Snowmelt_NDSI.csv"), row.names = FALSE)
              
              #Add dates of snowmelt to the plot 'p_Locations_NDSI_grid'
              p_Locations_NDSI_Snowmelt_grid <- p_Locations_NDSI_grid +
                geom_point(data=df_Snowmelt_NDSI_Locations[!is.na(df_Snowmelt_NDSI_Locations$doy),], aes(x=doy, y=NDSI_threshold), col="red", size=3)
              
              ggsave(plot=p_Locations_NDSI_Snowmelt_grid, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Plot_NDSI.pdf"), width=12, height = 10)
              }
              
      #VEGETATION - Fit a Generalized Additive Model (GAM) through the NDVI data for each location
         {
    
         #(a) Create an empty dataframe
             df_Locations_NDVI <- data.frame(NDVI=numeric(),
                                             Date=factor(),
                                             doy=numeric(),
                                             LocationID=factor(),
                                             outliers=logical())

             df_Locations_NDVI_predictions <- data.frame(LocationID=character(),
                                                         doy=numeric(),
                                                         NDVI_gam_predict=numeric(),
                                                         stringsAsFactors=FALSE)

         #(b) Loop through all Locations and fit a separate gam with sequential outlier removal to the Location specific NDVI data
             for(i in unique(Locations_Data_BandValues$LocationID)){

               #For debugging
               #i=unique(Locations_Data_BandValues$LocationID)[1]

               #Select Location-specific subset of data:
               df_Location_NDVI_new <- Locations_Data_BandValues[Locations_Data_BandValues$LocationID==i & !is.na(Locations_Data_BandValues$NDVI),
                                                   c("NDVI", "Date", "doy", "LocationID")]

               #Fit a gam through the location-specific NDVI ~ doy data and emply sequential outlier removal
               df_Location_NDVI_new <- f_gam_SeqRemOutliers(data=df_Location_NDVI_new, y="NDVI", x="doy", outlier_removal=outlier_removal,
                                               outlier_thresh_1=outlier_thresh_1, outlier_thresh_2=outlier_thresh_2,
                                               default_k=gam_k_outlier)

               #Sort df_Location_NDVI_new by doy
               df_Location_NDVI_new <- df_Location_NDVI_new[order(df_Location_NDVI_new$doy), ]

               #Bind the location-specific dataframe with GAM estimates to the dataframe containing all dataframes from previous iterations
               df_Locations_NDVI <- rbind(df_Locations_NDVI, df_Location_NDVI_new)

               #Create more detailed predictions (not only at the doy present in the datframe) at a 1-day interval to plot more smooth curves

                 #Refit GAM through data
                 index <- which(df_Location_NDVI_new$outliers==FALSE)
                 max_k_length2 <- length(unique(df_Location_NDVI_new[index, 'doy']))
                 mod_gam <- with(df_Location_NDVI_new[index,], mgcv::gam(NDVI ~ s(doy, k=min(gam_k, length(index)-1, max_k_length2-1)), method="REML"))
  
                 #Use gam to make predictions on a more detailed (1-day) day of year interval
                 df_Locations_NDVI_predictions_new <- data.frame(LocationID=i, doy=seq(min(df_Location_NDVI_new$doy), max(df_Location_NDVI_new$doy), 1))
                 df_Locations_NDVI_predictions_new$NDVI_gam_predict <- stats::predict(mod_gam, newdata=df_Locations_NDVI_predictions_new, type="response")
                 df_Locations_NDVI_predictions_new <- df_Locations_NDVI_predictions_new[order(df_Locations_NDVI_predictions_new$doy), ]
  
                 #Add predictions to df_Locations_NDVI_predictions dataframe:
                 df_Locations_NDVI_predictions <- rbind(df_Locations_NDVI_predictions, df_Locations_NDVI_predictions_new)

                 }

             #Change column LocationID to a factor:
             df_Locations_NDVI$LocationID <- as.factor(as.character(df_Locations_NDVI$LocationID))
             df_Locations_NDVI_predictions$LocationID <- as.factor(as.character(df_Locations_NDVI_predictions$LocationID))

             #Save dataframe with GAM fits for NDVI
             #write.csv(df_Locations_NDVI, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Data_NDVI_GAM.csv"), row.names = FALSE)
             write.csv(df_Locations_NDVI_predictions, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_GAM_Predictions_NDVI.csv"), row.names = FALSE)
             
         #(c) Plot the raw NDVI datapoints and gam predictions for each location:

             #Plot NDVI and model predictions for all locations in a single plot
             p_Locations_NDVI = ggplot()+
                geom_point(data=df_Locations_NDVI, aes(x=doy, y=NDVI, fill=LocationID, col=LocationID))+
                geom_line(data=df_Locations_NDVI_predictions, aes(x=doy, y=NDVI_gam_predict, col=LocationID)) +
                xlab("Day of year") +
                ylab("NDVI") +
                theme_tom()
               
             #Plot NDVI and model predictions in a separate plot per location
             p_Locations_NDVI_grid = ggplot()+
                geom_point(data=df_Locations_NDVI[df_Locations_NDVI$outliers==FALSE,], aes(x=doy, y=NDVI), col="black")+
                geom_point(data=df_Locations_NDVI[df_Locations_NDVI$outliers==TRUE,], aes(x=doy, y=NDVI), col="black", pch=16, alpha=0.2)+
                geom_line(data=df_Locations_NDVI_predictions, aes(x=doy, y=NDVI_gam_predict), col="#08a31a", lwd=1.25) +
                geom_vline(xintercept = 150, colour="grey", lty=2)+
                facet_wrap(~LocationID, ncol=ceiling(length(unique(df_Locations_NDVI$LocationID))^0.5))+
                xlab("Day of year") +
                ylab("NDVI") +
                theme_tom()

                ggsave(plot=p_Locations_NDVI_grid, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Plot_NDVI.pdf"), width=12, height = 10)
              }
         
      #MOISTURE - Fit a Generalized Additive Model (GAM) through the NDMI data for each location
         {
         #(a) Create an empty dataframe
             df_Locations_NDMI <- data.frame(NDMI=numeric(),
                                             Date=factor(),
                                             doy=numeric(),
                                             LocationID=factor(),
                                             outliers=logical())

             df_Locations_NDMI_predictions <- data.frame(LocationID=character(),
                                                         doy=numeric(),
                                                         NDMI_gam_predict=numeric(),
                                                         stringsAsFactors=FALSE)

         #(b) Loop through all locations and fit a separate gam with sequential outlier removal to the location specific NDMI data
             for(i in unique(Locations_Data_BandValues$LocationID)){

               #For debugging
               #i=unique(Locations_Data_BandValues$LocationID)[1]

               #Select location-specific subset of data:
               df_Location_NDMI_new <- Locations_Data_BandValues[Locations_Data_BandValues$LocationID==i & !is.na(Locations_Data_BandValues$NDMI),
                                                   c("NDMI", "Date", "doy", "LocationID")]

               #Fit a gam through the locations-specific NDMI ~ doy data and emply sequential outlier removal
               df_Location_NDMI_new <- f_gam_SeqRemOutliers(data=df_Location_NDMI_new, y="NDMI", x="doy", outlier_removal=outlier_removal,
                                                   outlier_thresh_1=outlier_thresh_1, outlier_thresh_2=outlier_thresh_2,
                                                   default_k=gam_k_outlier)

               #Sort df_Location_NDMI_new by doy
               df_Location_NDMI_new <- df_Location_NDMI_new[order(df_Location_NDMI_new$doy), ]

               #Bind the locations-specific dataframe with GAM estimates to the dataframe containing all dataframes from previous iterations
               df_Locations_NDMI <- rbind(df_Locations_NDMI, df_Location_NDMI_new)

               #Create more detailed predictions (not only at the doy present in the datframe) at a 1-day interval to plot more smooth curves

               #Refit GAM through data
               index <- which(df_Location_NDMI_new$outliers==FALSE)
               max_k_length3 <- length(unique(df_Location_NDMI_new[index, 'doy']))
               mod_gam <- with(df_Location_NDMI_new[index,], mgcv::gam(NDMI ~ s(doy, k=min(gam_k, length(index)-1, max_k_length3-1)), method="REML"))

               #Use gam to make predictions on a more detailed (1-day) day of year interval
               df_Locations_NDMI_predictions_new <- data.frame(LocationID=i, doy=seq(min(df_Location_NDMI_new$doy), max(df_Location_NDMI_new$doy), 1))
               df_Locations_NDMI_predictions_new$NDMI_gam_predict <- stats::predict(mod_gam, newdata=df_Locations_NDMI_predictions_new, type="response")
               df_Locations_NDMI_predictions_new <- df_Locations_NDMI_predictions_new[order(df_Locations_NDMI_predictions_new$doy), ]

               #Add predictions to df_Locations_NDMI_predictions dataframe:
               df_Locations_NDMI_predictions <- rbind(df_Locations_NDMI_predictions, df_Locations_NDMI_predictions_new)

               }

              #Change column LocationID to a factor:
               df_Locations_NDMI$LocationID <- as.factor(as.character(df_Locations_NDMI$LocationID))
               df_Locations_NDMI_predictions$LocationID <- as.factor(as.character(df_Locations_NDMI_predictions$LocationID))

              #Save dataframe with GAM fits for NDMI
               #write.csv(df_Locations_NDMI, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Data_NDMI_GAM.csv"), row.names = FALSE)
               write.csv(df_Locations_NDMI_predictions, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_GAM_Predictions_NDMI.csv"), row.names = FALSE)
               
         #(c) Plot the raw NDMI datapoints and gam predictions for each location:

             #Plot NDMI and model predictions for all locations in a single plot
              p_Locations_NDMI = ggplot()+
                geom_point(data=df_Locations_NDMI, aes(x=doy, y=NDMI, fill=LocationID, col=LocationID))+
                geom_line(data=df_Locations_NDMI_predictions, aes(x=doy, y=NDMI_gam_predict, col=LocationID)) +
                xlab("Day of year") +
                ylab("NDMI") +
                theme_tom()
               
             #Plot NDMI and model predictions in a separate plot per location
              p_Locations_NDMI_grid = ggplot()+
                geom_point(data=df_Locations_NDMI[df_Locations_NDMI$outliers==FALSE,], aes(x=doy, y=NDMI), col="black")+
                geom_point(data=df_Locations_NDMI[df_Locations_NDMI$outliers==TRUE,], aes(x=doy, y=NDMI), col="black", pch=16, alpha=0.2)+
                geom_line(data=df_Locations_NDMI_predictions, aes(x=doy, y=NDMI_gam_predict), col="#16acde", lwd=1.25)+
                geom_vline(xintercept = 150, colour="grey", lty=2)+
                facet_wrap(~LocationID, ncol=ceiling(length(unique(df_Locations_NDMI$LocationID))^0.5))+
                xlab("Day of year") +
                ylab("NDMI") +
                theme_tom()

                ggsave(plot=p_Locations_NDMI_grid, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Plot_NDMI.pdf"), width=12, height = 10)
              }
                
      #Plot NDSI, NDVI and NDMI together in a plot per location
         {
          #Create plot      
          p_Locations_BANDS_grid = ggplot()+
            geom_point(data=df_Locations_NDSI[df_Locations_NDSI$outliers==FALSE,], aes(x=doy, y=NDSI))+
            geom_point(data=df_Locations_NDSI[df_Locations_NDSI$outliers==TRUE,], aes(x=doy, y=NDSI), col="black", pch=16, alpha=0.2)+
            geom_point(data=df_Locations_NDVI[df_Locations_NDVI$outliers==FALSE,], aes(x=doy, y=NDVI), col="black")+
            geom_point(data=df_Locations_NDVI[df_Locations_NDVI$outliers==TRUE,], aes(x=doy, y=NDVI), col="black", pch=16, alpha=0.2)+
            geom_point(data=df_Locations_NDMI[df_Locations_NDMI$outliers==FALSE,], aes(x=doy, y=NDMI), col="black")+
            geom_point(data=df_Locations_NDMI[df_Locations_NDMI$outliers==TRUE,], aes(x=doy, y=NDMI), col="black", pch=16, alpha=0.2)+
            geom_line(data=df_Locations_NDSI_predictions, aes(x=doy, y=NDSI_gam_predict), col="#1620de", lwd=1.25)+
            geom_line(data=df_Locations_NDVI_predictions, aes(x=doy, y=NDVI_gam_predict), col="#08a31a", lwd=1.25)+
            geom_line(data=df_Locations_NDMI_predictions, aes(x=doy, y=NDMI_gam_predict), col="#16acde", lwd=1.25)+
            geom_vline(xintercept = 150, colour="grey", lty=2)+
            facet_wrap(~LocationID, ncol=ceiling(length(unique(df_Locations_NDMI$LocationID))^0.5))+
            xlab("Day of year") +
            ylab("Normalized Difference Band Index") +
            theme_tom()
  
            ggsave(plot=p_Locations_BANDS_grid, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Plot_AllBands.pdf"), width=14, height =12)
  
          }
      
    }

 #Method "SnowFraction"        
  if("snowfraction" %in% method){
   
    #(21): SNOW - Fit a Generalized Additive Model (GAM) through the SnowFraction data for each Location

        #We fit a Generalized Additive Model (GAM) through the data for each point location. We do this using a sequential
        #outlier-filtering process. We first fit a GAM through the data, calculating model predictions and residuals. We then
        #exclude all rows from the dataframe where the residual >= (0.4 * the range of the data). We then re-fit a GAM to
        #this reduced dataset, make predictions and calculate residuals. We then exclude all rows from the reduced dataframe
        #where the residual >= (0.2 * the range of the data). This gives us a final dataframe through which we fit a third
        #GAM and store its model predictions, given the dataset in which outliers were thus removed through two subsequent
        #steps. This whole process is executed using the function f_gam_SeqRemOutliers. Note that a sequential step is
        #required because initially some datapoints might falsely be assigned a large residual because of one extreme
        #outlier. After removal of this extreme outlier and refitting of the GAM, it can be better assessed which data
        #points truly have a large residual and can thus be assigned as 'true' outliers.
           f_gam_SeqRemOutliers <- f_gam_SeqRemOutliers #sourced

           #Specify sequential outlier thresholds:
            outlier_thresh_1=outlier_thresh_1
            outlier_thresh_2=outlier_thresh_2
            outlier_removal=outlier_removal

        #(a) Specify for which SnowFraction value the date of occurence will be calculated
            Snowfraction_threshold_vector = Snowfraction_threshold_vector

        #(b) Create an empty dataframe
            df_Locations_SnowFraction_GAM <- data.frame(NDSI_threshold=character(),
                                                        SnowFraction=numeric(),
                                                        Date=factor(),
                                                        doy=numeric(),
                                                        LocationID=factor(),
                                                        outliers=logical())

            df_Locations_SnowFraction_GAM_predictions <- data.frame(LocationID=character(),
                                                                    NDSI_threshold=character(),
                                                                    doy=numeric(),
                                                                    SnowFraction_gam_predict=numeric(),
                                                                    stringsAsFactors=FALSE)

        #(c) Loop through all Locations and fit a separate GAM with sequential outlier removal to the location specific SnowFraction data
            for(i in unique(df_Locations_SnowFraction$LocationID)){

              #For debugging
              #i=unique(df_Locations_SnowFraction$LocationID)[1]

              #Loop through all NDSI_thresholds
              for(j in unique(df_Locations_SnowFraction$NDSI_threshold)){
                
                #For debugging
                #j=unique(df_Locations_SnowFraction$NDSI_threshold)[1]
              
                #Select Location-specific subset of data:
                df_Location_SnowFraction_GAM_new <- df_Locations_SnowFraction[df_Locations_SnowFraction$LocationID==i & 
                                                                              df_Locations_SnowFraction$NDSI_threshold==j &
                                                                              !is.na(df_Locations_SnowFraction$SnowFraction),
                                                                              c("NDSI_threshold", "SnowFraction", "Date", "doy", "LocationID")]
                
                #Fit a gam through the Location-specific SnowFraction ~ doy data and employ sequential outlier removal
                df_Location_SnowFraction_GAM_new <- f_gam_SeqRemOutliers(data=df_Location_SnowFraction_GAM_new, y="SnowFraction", x="doy", outlier_removal=outlier_removal,
                                                                         outlier_thresh_1=outlier_thresh_1, outlier_thresh_2=outlier_thresh_2,
                                                                         default_k=gam_k_outlier)
                
                #Sort df_Location_SnowFraction_GAM_new by doy:
                df_Location_SnowFraction_GAM_new <- df_Location_SnowFraction_GAM_new[order(df_Location_SnowFraction_GAM_new$doy),]

                #Bind the Location-specific dataframe with GAM estimates to the dataframe containing all dataframes from previous iterations
                df_Locations_SnowFraction_GAM <- rbind(df_Locations_SnowFraction_GAM, df_Location_SnowFraction_GAM_new)

                #Create more detailed predictions (not only at the doy present in the datframe) at a 1-day interval to plot more smooth curves
  
                 #Refit GAM through data
                  index <- which(df_Location_SnowFraction_GAM_new$outliers==FALSE)
                  max_k_length4 <- length(unique(df_Location_SnowFraction_GAM_new[index, 'doy']))
                  mod_gam <- with(df_Location_SnowFraction_GAM_new[index,], mgcv::gam(SnowFraction ~ s(doy, k=min(gam_k, length(index)-1, max_k_length4-1)), method="REML"))
  
                 #Use gam to make predictions on a more detailed (1-day) day of year interval
                  df_Locations_SnowFraction_GAM_predictions_new <- data.frame(LocationID=i, NDSI_threshold=j, doy=seq(min(df_Location_SnowFraction_GAM_new$doy), max(df_Location_SnowFraction_GAM_new$doy), 1))
                  df_Locations_SnowFraction_GAM_predictions_new$SnowFraction_gam_predict <- stats::predict(mod_gam, newdata=df_Locations_SnowFraction_GAM_predictions_new, type="response")
                  df_Locations_SnowFraction_GAM_predictions_new <- df_Locations_SnowFraction_GAM_predictions_new[order(df_Locations_SnowFraction_GAM_predictions_new$doy),]
   
                 #Add predictions to df_Locations_SnowFraction_GAM_predictions dataframe:
                  df_Locations_SnowFraction_GAM_predictions <- rbind(df_Locations_SnowFraction_GAM_predictions, df_Locations_SnowFraction_GAM_predictions_new)

                }

              }
              
            #Change column LocationID to a factor:
             df_Locations_SnowFraction_GAM$LocationID <- as.factor(as.character(df_Locations_SnowFraction_GAM$LocationID))
             df_Locations_SnowFraction_GAM_predictions$LocationID <- as.factor(as.character(df_Locations_SnowFraction_GAM_predictions$LocationID))

             #Save dataframe with GAM fits for the SnowFraction data
             #write.csv(df_Locations_SnowFraction_GAM, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Data_SnowFraction_GAM.csv"), row.names = FALSE)
             write.csv(df_Locations_SnowFraction_GAM_predictions, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_GAM_Predictions_SnowFraction.csv"), row.names = FALSE)
             
        #(d) Plot the raw SnowFraction datapoints and gam predictions for each Location:

            #Plot SnowFraction and model predictions for all locations in a single plot
             p_Locations_SnowFraction = ggplot()+
               geom_point(data=df_Locations_SnowFraction_GAM, aes(x=doy, y=SnowFraction, fill=LocationID, col=LocationID))+
               geom_line(data=df_Locations_SnowFraction_GAM_predictions, aes(x=doy, y=SnowFraction_gam_predict, col=LocationID)) +
               xlab(paste0("Day of year (starting at 01-01-", year_ID,")")) +
               ylab("SnowFraction") +
               facet_wrap(~NDSI_threshold, ncol=3)+
               theme_tom()
              
            #Plot SnowFraction and model predictions in a separate plot per location
             p_Locations_SnowFraction_grid = ggplot()+
               geom_point(data=df_Locations_SnowFraction_GAM[df_Locations_SnowFraction_GAM$outliers==FALSE,], aes(x=doy, y=SnowFraction))+
               geom_point(data=df_Locations_SnowFraction_GAM[df_Locations_SnowFraction_GAM$outliers==TRUE,], aes(x=doy, y=SnowFraction), col="black", pch=16, alpha=0.2)+
               geom_line(data=df_Locations_SnowFraction_GAM_predictions, aes(x=doy, y=SnowFraction_gam_predict), col="#1620de", linewidth=1.25)+
               geom_vline(xintercept = 150, colour="grey", lty=2)+
               geom_hline(yintercept = Snowfraction_threshold_vector, colour="grey", lty=2)+
               facet_wrap(~LocationID + NDSI_threshold, ncol=ceiling(length(unique(df_Locations_SnowFraction_GAM$LocationID))^0.5))+
               xlab(paste0("Day of year (starting at 01-01-", year_ID,")")) +
               ylab("Snowcover fraction per location") +
               theme_tom()

               #ggsave(plot=p_Locations_SnowFraction_grid, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Plot_SnowFraction.pdf"), width=12, height = 10)

        #(e) Calculate at which day of year the SnowFraction is equal to Snowfraction_threshold_vector for each location using predictions from mod_gam

             #Create an empty dataframe
             df_Snowmelt_SnowFraction_Locations <- data.frame(NDSI_threshold=character(),
                                                     Snowfraction_threshold=character(),
                                                     doy=numeric(),
                                                     LocationID=character(),
                                                     stringsAsFactors=FALSE)

             #Setup parallel processing
             numCores <- detectCores()
             cl <- makePSOCKcluster(numCores)
             registerDoSNOW(cl)

            #Loop through all locations in the df_Locations_SnowFraction_GAM_predictions dataframe and extract doy for all Snowfraction levels:
            for(i in unique(df_Locations_SnowFraction_GAM_predictions$LocationID)){

               #For debugging
               #i=unique(df_Locations_SnowFraction_GAM_predictions$LocationID)[1] #for debugging
  
               #Select dataset with GAM predictions for current Location
               df_SnowFraction_gam <- df_Locations_SnowFraction_GAM_predictions[df_Locations_SnowFraction_GAM_predictions$LocationID==i & !is.na(df_Locations_SnowFraction_GAM_predictions$SnowFraction),]
  
               #Use the function f_detect_threshold_date_parallel to extract the moments the GAM predictions cross the thresholds in Snowfraction_threshold_vector
               results <- f_detect_threshold_date_parallel(subset=1, #data subset (not relevant here, set to 1)
                                                           pixelIDs_split = list(NDSI_threshold_vector), #levels of NDSI_threshold (input needs to be a list)
                                                           df_pixel_y = df_SnowFraction_gam, #dataframe containing GAM predictions
                                                           pixel_ID_column="NDSI_threshold", #Grouping column
                                                           y="SnowFraction_gam_predict", #response variable in GAM
                                                           x="doy", #predictor variable in GAM
                                                           pixel_gam_plots = FALSE, #Should GAM plots be created
                                                           y_threshold = Snowfraction_threshold_vector) #Which threshold values for 'y' should be calculated
  
               #Store dates of snowmelt per Location
               df_Snowmelt_SnowFraction_Locations_new <- results[[1]]
               df_Snowmelt_SnowFraction_Locations_new <- as.data.frame(do.call(rbind, df_Snowmelt_SnowFraction_Locations_new))
               colnames(df_Snowmelt_SnowFraction_Locations_new)[colnames(df_Snowmelt_SnowFraction_Locations_new)=="pixel_ID"] <- "NDSI_threshold"
               colnames(df_Snowmelt_SnowFraction_Locations_new)[colnames(df_Snowmelt_SnowFraction_Locations_new)=="x_threshold"] <- "doy"
               colnames(df_Snowmelt_SnowFraction_Locations_new)[colnames(df_Snowmelt_SnowFraction_Locations_new)=="y_threshold"] <- "Snowfraction_threshold"
               df_Snowmelt_SnowFraction_Locations_new$LocationID <- i
  
               #Add snowmelt data for LocationID to general dataframe
               df_Snowmelt_SnowFraction_Locations <- rbind(df_Snowmelt_SnowFraction_Locations, df_Snowmelt_SnowFraction_Locations_new)
  
              }

            #Turn parallel processing off and run sequentially again after this point
            stopCluster(cl)
            registerDoSEQ()

            #Add coordinates to each location
            df_Snowmelt_SnowFraction_Locations <- left_join(df_Snowmelt_SnowFraction_Locations, Locations_sf, by=c("LocationID"))

            #Save dates of snowmelt per Location per SnowFraction threshold as a .csv file
            write.csv(df_Snowmelt_SnowFraction_Locations, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Snowmelt_Snowfraction.csv"), row.names = FALSE)

            #Add dates of snowmelt to the plot 'p_Locations_NDSI_grid'
            p_Locations_SnowFraction_Snowmelt_grid <- p_Locations_SnowFraction_grid +
              geom_point(data=df_Snowmelt_SnowFraction_Locations[!is.na(df_Snowmelt_SnowFraction_Locations$doy),], aes(x=doy, y=Snowfraction_threshold), col="red", size=3)

            ggsave(plot=p_Locations_SnowFraction_Snowmelt_grid, paste0(here(), "/Output/S2/08_Shapefile_Points_Snowmelt/", timestamp, "_", data_ID, "_Buffer", Buffer_radius_m, "_Res", resolution, "_Locations_Plot_Snowfraction.pdf"), width=12, height = 10)
          
        } 

##########################################################################################################################################################################

#The End
      
  #Print concluding remarks  
  {cat("\n") ; cat("\n")
    print("--------------------------------------------------------------------------------------------------------------------------")
    print(paste0("THE ANALYSIS HAS COMPLETED"))
    cat("\n")
    print(paste0("-Generated data and plots can be found at ", here(), "/Output/S2/08_Shapefile_Points_Snowmelt/"))
    print("--------------------------------------------------------------------------------------------------------------------------")
    cat("\n")}

##########################################################################################################################################################################
